no additional exception if test fails in setup

+++ src/Addin.Transform.Tests/HtmlTransformerTests.cs
@@ -30,11 +30,14 @@ namespace Addin.Transform.Tests
 		[TearDown]
 		public void TearDown()
 		{
-			if (File.Exists(_addin.PathToOutput))
+			if (_addin != null && File.Exists(_addin.PathToOutput))
 			{
 				File.Delete(_addin.PathToOutput);
 			}
-			_lexEntryRepository.Dispose();
+			if (_lexEntryRepository != null)
+			{
+				_lexEntryRepository.Dispose();
+			}
 			File.Delete(_filePath);
 			File.Delete(_liftFilePath);
 		}
