+++ src/LexicalTools.Tests/EntryViewControlTests.cs
@@ -94,8 +94,14 @@ namespace WeSay.LexicalTools.Tests
 		[TearDown]
 		public void TearDown()
 		{
-			_lexEntryRepository.Dispose();
-			_tempFolder.Delete();
+			if (_lexEntryRepository != null)
+			{
+				_lexEntryRepository.Dispose();
+			}
+			if (_tempFolder != null)
+			{
+				_tempFolder.Delete();
+			}
 		}
 
 		[Test]
